requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.1.0
celery==5.3.4
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-writable-nested==0.7.0
drf-yasg==1.21.7
idna==3.7
importlib-metadata==6.8.0
importlib_resources==6.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.2
Markdown==3.4.4
MarkupSafe==2.1.5
numpy==1.25.2
openapi-codec==1.3.2
packaging==24.1
pandas==2.2.2
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
redis==5.0.7
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.4
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.2
vine==5.0.0
wcwidth==0.2.8
zipp==3.16.2
